+++ runzi/automation/local_config.py
@@ -14,7 +14,7 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-from nshm_toshi_client import API_URL, COGNITO_DOMAIN, M2M_SECRET_ARN, S3_URL, get_auth_kwargs  # noqa: F401, E402
+from nshm_toshi_client import API_URL, S3_URL, get_auth_kwargs  # noqa: F401, E402
 
 logging.basicConfig(level=logging.INFO)
 log = logging.getLogger(__name__)